Actualizar la firma de la función MoveOmnidirectionalBase para usar un tipo de dato double en el parámetro offset y ajustar el valor de kCorrectionDegreeOffset a 16 en demoStriker.cpp

+++ lib/Motors/motors.cpp
@@ -27,7 +27,7 @@ void Motors::StopAllMotors()
     upper_right_motor_.StopMotor();
 };
 
-void Motors::MoveOmnidirectionalBase(double degree, float speed, double speed_w, int offset)
+void Motors::MoveOmnidirectionalBase(double degree, float speed, double speed_w, double offset)
 {
     if (degree > 0){
         degree -= offset;
